sui_adapter_v1::programmable_transactions

Module linkage_view

Structs

LinkageInfo

LinkageView

Exposes module and linkage resolution to the Move runtime. The first by delegating to resolver and the second via linkage information that is loaded from a move package.

SavedLinkage