sui_adapter_v1::temporary_store
Struct TemporaryStore

pub struct TemporaryStore<'backing> { /* private fields */ }

Implementations

impl<'backing> TemporaryStore<'backing>

pub fn new( store: &'backing dyn BackingStore, input_objects: InputObjects, receiving_objects: Vec<ObjectRef>, tx_digest: TransactionDigest, protocol_config: &ProtocolConfig, ) -> Self

Creates a new store associated with an authority store, and populates it with initial objects.

pub fn objects(&self) -> &BTreeMap<ObjectID, Object>

pub fn update_object_version_and_prev_tx(&mut self)

pub fn into_inner(self) -> InnerTemporaryStore

Break up the structure and return its internal stores (objects, active_inputs, written, deleted)

pub fn into_effects( self, shared_object_refs: Vec<SharedInput>, transaction_digest: &TransactionDigest, transaction_dependencies: BTreeSet<TransactionDigest>, gas_cost_summary: GasCostSummary, status: ExecutionStatus, gas_charger: &mut GasCharger, epoch: EpochId, ) -> (InnerTemporaryStore, TransactionEffects)

pub fn mutate_input_object(&mut self, object: Object)

Mutate a mutable input object. This is used to mutate input objects outside of PT execution.

pub fn mutate_child_object(&mut self, old_object: Object, new_object: Object)

Mutate a child object outside of PT. This should be used extremely rarely. Currently it’s only used by advance_epoch_safe_mode because it’s all native without PT. This should almost never be used otherwise.

pub fn upgrade_system_package(&mut self, package: Object)

Upgrade system package during epoch change. This requires special treatment since the system package to be upgraded is not in the input objects. We could probably fix above to make it less special. Due to the special treatment, we need to read from object store explicitly to obtain the modified_at information.

pub fn create_object(&mut self, object: Object)

Crate a new objcet. This is used to create objects outside of PT execution.

pub fn delete_input_object(&mut self, id: &ObjectID)

Delete a mutable input object. This is used to delete input objects outside of PT execution.

pub fn drop_writes(&mut self)

pub fn read_object(&self, id: &ObjectID) -> Option<&Object>

pub fn save_loaded_runtime_objects( &mut self, loaded_runtime_objects: BTreeMap<ObjectID, DynamicallyLoadedObjectMetadata>, )

pub fn estimate_effects_size_upperbound(&self) -> usize

pub fn written_objects_size(&self) -> usize

pub fn conserve_unmetered_storage_rebate( &mut self, unmetered_storage_rebate: u64, )

If there are unmetered storage rebate (due to system transaction), we put them into the storage rebate of 0x5 object. TODO: This will not work for potential future new system transactions if 0x5 is not in the input. We should fix this.

impl TemporaryStore<'_>

pub fn check_ownership_invariants( &self, sender: &SuiAddress, gas_charger: &mut GasCharger, is_epoch_change: bool, ) -> SuiResult<()>

impl TemporaryStore<'_>

pub fn check_execution_results_consistency(&self) -> Result<(), ExecutionError>

impl TemporaryStore<'_>

pub fn advance_epoch_safe_mode( &mut self, params: &AdvanceEpochParams, protocol_config: &ProtocolConfig, )

impl TemporaryStore<'_>

pub fn check_sui_conserved( &self, simple_conservation_checks: bool, gas_summary: &GasCostSummary, ) -> Result<(), ExecutionError>

Check that this transaction neither creates nor destroys SUI. This should hold for all txes except the epoch change tx, which mints staking rewards equal to the gas fees burned in the previous epoch. Specifically, this checks two key invariants about storage fees and storage rebate:

all SUI in storage rebate fields of input objects should flow either to the transaction storage rebate, or the transaction non-refundable storage rebate

all SUI charged for storage should flow into the storage rebate field of some output object

This function is intended to be called after we have charged for gas + applied the storage rebate to the gas object, but before we have updated object versions.

pub fn check_sui_conserved_expensive( &self, gas_summary: &GasCostSummary, advance_epoch_gas_summary: Option<(u64, u64)>, layout_resolver: &mut impl LayoutResolver, ) -> Result<(), ExecutionError>

Check that this transaction neither creates nor destroys SUI. This more expensive check will check a third invariant on top of the 2 performed by check_sui_conserved above:

all SUI in input objects (including coins etc in the Move part of an object) should flow either to an output object, or be burned as part of computation fees or non-refundable storage rebate

This function is intended to be called after we have charged for gas + applied the storage rebate to the gas object, but before we have updated object versions. The advance epoch transaction would mint epoch_fees amount of SUI, and burn epoch_rebates amount of SUI. We need these information for this check.